+++ internal/service/ec2/vpc_security_group_ingress_rule_test.go
@@ -10,9 +10,8 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/google/go-cmp/cmp"
-	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/path"
-	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	sdkacctest "github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
@@ -27,9 +26,9 @@ func TestNormalizeIPProtocol(t *testing.T) {
 	t.Parallel()
 
 	type testCase struct {
-		plannedValue  attr.Value
-		currentValue  attr.Value
-		expectedValue attr.Value
+		plannedValue  types.String
+		currentValue  types.String
+		expectedValue types.String
 		expectError   bool
 	}
 	tests := map[string]testCase{
@@ -54,15 +53,15 @@ func TestNormalizeIPProtocol(t *testing.T) {
 		name, test := name, test
 		t.Run(name, func(t *testing.T) {
 			ctx := context.Background()
-			request := tfsdk.ModifyAttributePlanRequest{
-				AttributePath:  path.Root("test"),
-				AttributePlan:  test.plannedValue,
-				AttributeState: test.currentValue,
+			request := planmodifier.StringRequest{
+				Path:       path.Root("test"),
+				PlanValue:  test.plannedValue,
+				StateValue: test.currentValue,
 			}
-			response := tfsdk.ModifyAttributePlanResponse{
-				AttributePlan: request.AttributePlan,
+			response := planmodifier.StringResponse{
+				PlanValue: request.PlanValue,
 			}
-			tfec2.NormalizeIPProtocol().Modify(ctx, request, &response)
+			tfec2.NormalizeIPProtocol().PlanModifyString(ctx, request, &response)
 
 			if !response.Diagnostics.HasError() && test.expectError {
 				t.Fatal("expected error, got no error")
@@ -72,7 +71,7 @@ func TestNormalizeIPProtocol(t *testing.T) {
 				t.Fatalf("got unexpected error: %s", response.Diagnostics)
 			}
 
-			if diff := cmp.Diff(response.AttributePlan, test.expectedValue); diff != "" {
+			if diff := cmp.Diff(response.PlanValue, test.expectedValue); diff != "" {
 				t.Errorf("unexpected diff (+wanted, -got): %s", diff)
 			}
 		})