+++ internal/service/meta/service_data_source.go
@@ -7,8 +7,7 @@ import (
 
 	"github.com/aws/aws-sdk-go/aws/endpoints"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	"github.com/hashicorp/terraform-plugin-framework/diag"
-	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
+	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-provider-aws/internal/framework"
 	"github.com/hashicorp/terraform-provider-aws/internal/slices"
@@ -33,52 +32,42 @@ func (d *dataSourceService) Metadata(_ context.Context, request datasource.Metad
 	response.TypeName = "aws_service"
 }
 
-// GetSchema returns the schema for this data source.
-func (d *dataSourceService) GetSchema(context.Context) (tfsdk.Schema, diag.Diagnostics) {
-	schema := tfsdk.Schema{
-		Attributes: map[string]tfsdk.Attribute{
-			"dns_name": {
-				Type:     types.StringType,
+// Schema returns the schema for this data source.
+func (d *dataSourceService) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		Attributes: map[string]schema.Attribute{
+			"dns_name": schema.StringAttribute{
 				Optional: true,
 				Computed: true,
 			},
-			"id": {
-				Type:     types.StringType,
+			"id": schema.StringAttribute{
 				Optional: true,
 				Computed: true,
 			},
-			"partition": {
-				Type:     types.StringType,
+			"partition": schema.StringAttribute{
 				Computed: true,
 			},
-			"region": {
-				Type:     types.StringType,
+			"region": schema.StringAttribute{
 				Optional: true,
 				Computed: true,
 			},
-			"reverse_dns_name": {
-				Type:     types.StringType,
+			"reverse_dns_name": schema.StringAttribute{
 				Optional: true,
 				Computed: true,
 			},
-			"reverse_dns_prefix": {
-				Type:     types.StringType,
+			"reverse_dns_prefix": schema.StringAttribute{
 				Optional: true,
 				Computed: true,
 			},
-			"service_id": {
-				Type:     types.StringType,
+			"service_id": schema.StringAttribute{
 				Optional: true,
 				Computed: true,
 			},
-			"supported": {
-				Type:     types.BoolType,
+			"supported": schema.BoolAttribute{
 				Computed: true,
 			},
 		},
 	}
-
-	return schema, nil
 }
 
 // Read is called when the provider must read data source values in order to update state.