+++ internal/service/resourceexplorer2/exports_test.go
@@ -1,0 +1,7 @@
+package resourceexplorer2
+
+// Exports for use in tests only.
+var (
+	FindIndex     = findIndex
+	ResourceIndex = newResourceIndex
+)